pyproject.toml
[project]
name = "sarvam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "pyaudio>=0.2.14",
    "requests>=2.32.4",
    "requests-toolbelt>=1.0.0",
]
